prplmesh-hostapd: added disable_all_ap_vaps()
added functionality to disable all vaps that
are access-points (as opposed to STA for example).
the code mimics the existing functionality taken from
ap_wlan_hal_dwpal.cpp.
added funcvtionlity and unit test
note:
the example unit-test configuration file did not
contain any vap with the mode= line, therefore it
was added few of this.
also:
it was discovered that it is OK to look for non-exiting key
within vap's configuration and the code was changed accordingly
https://jira.prplfoundation.org/browse/PPM-249

Signed-off-by: Ran Regev <ran.regev@devalore.com>