transport: move event loop to main function
https://jira.prplfoundation.org/browse/PPM-18

This is a preparatory commit.

Application event loop is currently defined in Ieee1905Transport::run().
Since event loop is going to be a dependency for the interface state
monitor as well as for the transport class, move it to the main
function so it can later be reused.

The event loop must be then injected as a parameter in constructor to
all classes that depend on it like, for example, the BrokerServer class.
While we are at it, fix a bug in the constructor of BrokerServer (it is
creating a shared_ptr from a reference) and another one in destructor
(it must remove any registered event handler).

Since event loop is now a member variable of the Ieee1905Transport class,
remove the methods add_event, del_event and run which are not necessary
any more.

Also update unit tests accordingly.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>