bcl: remove template parameter from event loop
https://jira.prplfoundation.org/browse/PPM-18

This is a preparatory commit.

EventLoop interface and implementation are not intended to deal with
sockets only (or any other specific class). Instead, EventLoop will
deal with OS resources implementing the file descriptor interface. So
there is no need to couple EventLoop whith any specific class, even if
using a template for that.

In order to use the current event loop with the new sockets API to
implement the interface state monitor, event loop interface and
implementation have to be modified to use `int` type to refer to the
source of events, being that integer the file descriptor of the OS
resource where events occur. In the future, the file descriptor will be
represented by a `FileDescriptor` class which, among other features,
will close the file descriptor in its destructor. Current `Socket`
class will be eventually deprecated in the future.

Method to register handlers for events on a given file descriptor has
been renamed from `add_event` to the more appropriate name
`register_handlers`. For the same reason, the method to remove those
handlers when they are not needed any more has been renamed from
`del_event` to `remove_handlers`.

`SocketEventLoop` class has been renamed to `EventLoopImpl` and the
unit tests modified accordingly.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>