db: initial-radio is set only if stay-on-initial-radio enabled
The initial-radio is an internal persistent parameter, used to provide
the functionality of stay-on-initial-radio across reboots.
Since it is an internal parameter (although persistent) it should not
cause a timestamp update. It should also be available only if the
stay-on-initial-radio is enabled.
When client connects, if the stay-on-initial-radio is enabled and the
initial-radio is not configured, the initial-radio is set to the parent
radio mac (not bssid).
When stay-on-initial-radio is disabled, the initial-radio is cleared
(even if client is currently connected) to allow re-aquiring the
initial-radio when the stay-on-initial-radio is enabled.

Signed-off-by: Adam Dov <adamx.dov@intel.com>